Add files via upload
Welcome to our C# Playground! 🎮 Dive into the fundamentals of C# with our collection of exploratory scripts and test cases. Whether you're a beginner honing your syntax skills or an experienced developer brushing up on the basics, this repository offers a hands-on environment to experiment, learn, and master the essentials of C# programming!